Backport RemoveDeletedSLContentishChildren DataCollector for ftw.contentpage
