+++ ftw/publisher/sender/workflows/subscribers.py
@@ -1,5 +1,6 @@
 from Acquisition import aq_inner
 from Acquisition import aq_parent
+from ftw.publisher.core.adapters.simplelayout_utils import is_sl_contentish
 from ftw.publisher.sender.utils import is_temporary
 from ftw.publisher.sender.workflows.interfaces import DELETE_ACTIONS
 from ftw.publisher.sender.workflows.interfaces import IPublisherContextState
@@ -8,16 +9,6 @@ from ftw.publisher.sender.workflows.interfaces import PUSH_ACTIONS
 from Products.CMFPlone.interfaces import IPloneSiteRoot
 from zope.component import getMultiAdapter
 from zope.component import getUtility
-import pkg_resources
-
-
-try:
-    pkg_resources.get_distribution('ftw.simplelayout')
-except pkg_resources.DistributionNotFound:
-    FTW_SIMPLELAYOUT_AVAILABLE = False
-else:
-    FTW_SIMPLELAYOUT_AVAILABLE = True
-    from ftw.publisher.core.adapters.ftw_simplelayout import is_sl_contentish
 
 
 _marker = '_publisher_event_already_handled'
@@ -63,18 +54,16 @@ def publish_after_transition(context, event):
         context.restrictedTraverse('@@publisher.delete')()
 
 
-
 def handle_remove_event(context, event):
     """
     Before a object is remvoed the event handler crates a remove job.
     """
-
     # the event is notified for every subobject, but we only want to check
     # the top object which the users tries to delete
     if context is not event.object:
         return
 
-    if FTW_SIMPLELAYOUT_AVAILABLE and is_sl_contentish(context):
+    if is_sl_contentish(context):
         # Do not delete sl contentish objects, they are deleted when the
         # associated sl container is published.
         return
